fucking around with fft scaling - mips etc are fine up until 4k/4k then shit falls apart. nans in the sum image and the mip skyrockets. I dk wtf is going on. I

+++ src/gpu/src/Histogram.cu
@@ -30,7 +30,7 @@ __global__ void histogram_smem_atomics(const Npp32f* in, int4 dims, float *out, 
 
 
   // initialize temporary accumulation array in shared memory
-  extern __shared__ unsigned int smem[];
+  extern __shared__  int smem[];
   for (int i = t; i < n_bins; i += nt) smem[i] = 0;
   __syncthreads();
 
@@ -143,7 +143,7 @@ void Histogram::BufferInit(NppiSize npp_ROI)
 	 gridDims_accum_array = dim3((histogram_n_bins + threadsPerBlock_accum_array.x - 1) / threadsPerBlock_accum_array.x,1,1);
 
 
-	 size_of_temp_hist = (gridDims_img.x*gridDims_img.y*histogram_n_bins*sizeof(unsigned int));
+	 size_of_temp_hist = (gridDims_img.x*gridDims_img.y*histogram_n_bins*sizeof(int));
 
 
 	 // Array of temporary storage to accumulate the shared mem to
@@ -170,7 +170,7 @@ void Histogram::AddToHistogram(GpuImage &input_image)
 
 
 	pre_checkErrorsAndTimingWithSynchronization(cudaStreamPerThread);
-	histogram_smem_atomics<<< gridDims_img,threadsPerBlock_img, (histogram_n_bins)*sizeof(unsigned int), input_image.nppStream.hStream>>>((const Npp32f*)input_image.real_values_gpu, input_image.dims, histogram, histogram_n_bins,histogram_min,histogram_step,max_padding);
+	histogram_smem_atomics<<< gridDims_img,threadsPerBlock_img, (histogram_n_bins)*sizeof(int), input_image.nppStream.hStream>>>((const Npp32f*)input_image.real_values_gpu, input_image.dims, histogram, histogram_n_bins,histogram_min,histogram_step,max_padding);
 	checkErrorsAndTimingWithSynchronization(cudaStreamPerThread);
 
 //	pre_checkErrorsAndTimingWithSynchronization(cudaStreamPerThread);
@@ -194,8 +194,8 @@ void Histogram::CopyToHostAndAdd(long* array_to_add_to)
 
 	// Make a temporary copy of the cummulative histogram on the host and then add on the host. TODO errorchecking
 	float* tmp_array;
-	checkCudaErrors(cudaMallocHost(&tmp_array, histogram_n_bins*sizeof(long)));
-	checkCudaErrors(cudaMemcpy(tmp_array, this->cummulative_histogram,histogram_n_bins*sizeof(long),cudaMemcpyDeviceToHost));
+	checkCudaErrors(cudaMallocHost(&tmp_array, histogram_n_bins*sizeof(float)));
+	checkCudaErrors(cudaMemcpy(tmp_array, this->cummulative_histogram,histogram_n_bins*sizeof(float),cudaMemcpyDeviceToHost));
 
 	for (int iBin = 0; iBin < histogram_n_bins; iBin++)
 	{
